+++ projects/ng-base/base-components/components/change-password/change-password.component.ts
@@ -40,7 +40,7 @@ export class ChangePasswordComponent implements OnInit {
         password: new FormControl<string>('', [Validators.required, securePasswordValidator()]),
         passwordConfirm: new FormControl<string>('', [Validators.required, securePasswordValidator()]),
       },
-      { validators: Validation.match('password', 'passwordConfirm') }
+      Validation.match('password', 'passwordConfirm') as any
     );
   }