+++ projects/ng-base/src/guards/can-deactivate.guard.ts
@@ -1,12 +1,12 @@
 import { Injectable } from '@angular/core';
-import { CanDeactivate } from '@angular/router';
+
 import { ICanComponentDeactivate } from '@lenne.tech/ng-base/shared';
 
 /**
  * Guard to check if a component can be deactivated
  */
 @Injectable()
-export class CanDeactivateGuard implements CanDeactivate<ICanComponentDeactivate> {
+export class CanDeactivateGuard {
   canDeactivate(component: ICanComponentDeactivate) {
     return component.canDeactivate ? component.canDeactivate() : true;
   }